enable tagging the entire scope with patch/minor/major
