+++ src/api/consumer/lib/tag.ts
@@ -1,6 +1,5 @@
 import R from 'ramda';
 import semver from 'semver';
-
 import { Analytics } from '../../../analytics/analytics';
 import { BitId, BitIds } from '../../../bit-id';
 import { POST_TAG_ALL_HOOK, POST_TAG_HOOK, PRE_TAG_ALL_HOOK, PRE_TAG_HOOK } from '../../../constants';
@@ -47,7 +46,7 @@ type TagParams = {
   ignoreNewestVersion: boolean;
   id: string;
   all: boolean;
-  scope?: string;
+  scope?: string | boolean;
   includeImported: boolean;
 } & BasicTagParams;
 
@@ -122,7 +121,7 @@ async function getComponentsToTag(
   isAll: boolean,
   id?: string
 ): Promise<{ bitIds: BitId[]; warnings: string[] }> {
-  const warnings = [];
+  const warnings: string[] = [];
   const idHasWildcard = id && hasWildcard(id);
   if (id && !idHasWildcard) {
     const bitId = consumer.getParsedId(id);
@@ -142,12 +141,10 @@ async function getComponentsToTag(
     ? await componentsList.listTagPendingOfAllScope(includeImported)
     : await componentsList.listTagPendingComponents();
 
-  if (isAllScope) {
+  if (isAllScope && exactVersion) {
     const tagPendingComponentsLatest = await consumer.scope.latestVersions(tagPendingComponents, false);
     tagPendingComponentsLatest.forEach((componentId) => {
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      if (semver.gt(componentId.version!, exactVersion as string)) {
-        // @ts-ignore AUTO-ADDED-AFTER-MIGRATION-PLEASE-FIX!
+      if (componentId.version && semver.gt(componentId.version, exactVersion)) {
         warnings.push(`warning: ${componentId.toString()} has a version greater than ${exactVersion}`);
       }
     });
